requirements.txt
Django==5.1.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
requests==2.32.3
coverage==7.6.1
pylint==3.2.7
pylint-django==2.5.5
bandit==1.7.9